chore(resolver): split query.rs into directory module (488+95 lines)
